break up cluster context
