+++ pkg/cluster/internal/delete/delete.go
@@ -1,0 +1,50 @@
+/*
+Copyright 2019 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package delete
+
+import (
+	"fmt"
+	"os"
+	"strings"
+
+	"github.com/pkg/errors"
+	log "github.com/sirupsen/logrus"
+
+	"sigs.k8s.io/kind/pkg/cluster/internal/context"
+	"sigs.k8s.io/kind/pkg/cluster/nodes"
+)
+
+// Cluster deletes the cluster identified by ctx
+func Cluster(c *context.Context) error {
+	n, err := c.ListNodes()
+	if err != nil {
+		return errors.Wrap(err, "error listing nodes")
+	}
+
+	// try to remove the kind kube config file generated by "kind create cluster"
+	err = os.Remove(c.KubeConfigPath())
+	if err != nil && !os.IsNotExist(err) {
+		log.Warningf("Tried to remove %s but received error: %s\n", c.KubeConfigPath(), err)
+	}
+
+	// check if $KUBECONFIG is set and let the user know to unset if so
+	if strings.Contains(os.Getenv("KUBECONFIG"), c.KubeConfigPath()) {
+		fmt.Printf("$KUBECONFIG is still set to use %s even though that file has been deleted, remember to unset it\n", c.KubeConfigPath())
+	}
+
+	return nodes.Delete(n...)
+}
